百度网盘登陆处理

+++ base/src/main/java/cn/kk/base/io/model/netdisk/baidu/NetDisk_BaiduAccessTokenResponse.java
@@ -3,7 +3,7 @@ package cn.kk.base.io.model.netdisk.baidu;
 public class NetDisk_BaiduAccessTokenResponse {
 
 
-    public int expires_in;
+    public long expires_in;
     public String refresh_token;
     public String access_token;
     public String session_secret;
